+++ src/ros/rover/arm/arm_kinematics/arm_kinematics/include/arm_kinematics/inverse/inverse_kinematics_plugin.hpp
@@ -48,7 +48,7 @@ public:
    * \returns True if a solution was found
    */
   virtual bool get_position_ik(
-    const Eigen::Isometry3d & ik_pose,
+    const Eigen::Isometry3f & ik_pose,
     const std::vector<double> & ik_seed_state,
     std::vector<double> & solution_state) const = 0;
 
@@ -74,7 +74,7 @@ public:
    */
   virtual bool get_velocity_ik(
     const Eigen::Matrix<double, 6, 1> & ik_twist,
-    const Eigen::Isometry3d & ik_seed_pose,
+    const Eigen::Isometry3f & ik_seed_pose,
     const std::vector<double> & ik_seed_state,
     std::vector<double> & solution_velocities,
     double time_step) const;
